rxjava - folder crud events
